move chart to strategies/common
