+++ src/components/strategies/common/d3Chart/Candlesticks.tsx
