+++ src/components/strategies/common/d3Chart/D3ChartHandleLine.tsx
