+++ src/components/strategies/common/d3Chart/index.ts
