+++ src/components/strategies/common/d3Chart/overlapping/D3ChartOverlappingHandle.tsx
@@ -1,29 +1,26 @@
 import { D3ChartSettings, drag, Selection } from 'libs/d3';
-import { D3ChartHandleLine } from 'components/simulator/input/d3Chart/D3ChartHandleLine';
+import { D3ChartHandleLine } from 'components/strategies/common/d3Chart/D3ChartHandleLine';
 import {
   getSelector,
   useSelectable,
-} from 'components/simulator/input/d3Chart/utils';
+} from 'components/strategies/common/d3Chart/utils';
 import { useEffect } from 'react';
 
 interface Props {
   selector: string;
-  selectorOpposite: string;
   dms: D3ChartSettings;
   onDragStart?: (y: number) => void;
-  onDrag: (y: number) => void;
-  onDragEnd?: (y?: number, y2?: number) => void;
+  onDrag?: (y: number) => void;
+  onDragEnd?: (y: number) => void;
   color: string;
   label?: string;
   readonly?: boolean;
-  isLimit: boolean;
 }
 
-export const D3ChartHandle = ({
+export const D3ChartOverlappingHandle = ({
   onDragStart,
   onDrag,
   onDragEnd,
-  isLimit,
   ...props
 }: Props) => {
   const isSelectable = useSelectable(props.selector);
@@ -36,35 +33,20 @@ export const D3ChartHandle = ({
         y: Number(selection.select('line').attr('y1')),
       }))
       .on('start', ({ y }) => onDragStart?.(y))
-      .on('drag', ({ y }) => onDrag(y))
-      .on('end', ({ y }) => {
-        if (isLimit) {
-          onDragEnd?.(y, y);
-          return;
-        }
+      .on('drag', ({ y }) => onDrag?.(y))
+      .on('end', ({ y }) => onDragEnd?.(y));
 
-        const oppositeY = Number(
-          getSelector(props.selectorOpposite).select('line').attr('y1')
-        );
-
-        if (oppositeY < y) {
-          onDragEnd?.(y, oppositeY);
-        } else {
-          onDragEnd?.(oppositeY, y);
-        }
-      });
     if (!isSelectable) return;
+
     handleDrag(selection as Selection<Element, unknown, any, any>);
   }, [
     isSelectable,
-    onDragStart,
     onDrag,
     onDragEnd,
-    props.selector,
+    onDragStart,
     props.readonly,
-    props.selectorOpposite,
-    isLimit,
+    props.selector,
   ]);
 
-  return <D3ChartHandleLine {...props} isDraggable />;
+  return <D3ChartHandleLine {...props} isDraggable={!props.readonly} />;
 };
