+++ src/components/strategies/common/d3Chart/recurring/D3ChartRectDraggable.tsx
@@ -1,9 +1,9 @@
-import { D3ChartRect } from 'components/simulator/input/d3Chart/D3ChartRect';
+import { D3ChartRect } from 'components/strategies/common/d3Chart/D3ChartRect';
 import { drag, Selection } from 'd3';
 import {
   getSelector,
   useSelectable,
-} from 'components/simulator/input/d3Chart/utils';
+} from 'components/strategies/common/d3Chart/utils';
 import { D3ChartSettings } from 'libs/d3/types';
 import { useEffect, useState } from 'react';
