+++ src/components/strategies/common/d3Chart/xAxis.tsx
